feat: add `pgvector` extension
AI is a hot topic in Cloud Native environments (not only there).
Providing a way to easily add `pgvector` to CloudNativePG deployments
can potentially increase the adoption of Postgres in these contexts.

It is an extension that can be fully controlled by a standard Postgres
user, like the application user in CloudNativePG, without requiring
any settings in shared_preload_libraries.

License-wise that extension creates no problems.

Closes #70

Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>