ci: defer creating major `version` tag until release (#108)
Previously, we were always setting the `version` tag (e.g., 16, 17) when building
a container image. However, version 17 is still in beta and was tagged with `17`,
which could cause confusion. Users might mistake it for a stable release,
and tools like Renovate might attempt to update to this version.

This change ensures that the major `version` tag is not created until the version
is officially released.

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>